Bug: Remove Extra Definition of get_result_model
Removes a definition of get_result_model inside the definition of the
mixin AbstractPattern. It was a leftover from when get_origin_model
needed to be defined (now AbstractPattern is hard-coded to only take
Word's, which I suppose is hardly better practice). This commit should
really not be on this branch, but would then have to be merged with this
branch which would be a huge hassle and I am the only one working on
this project and nothing will ever come of it probably so whatever.