Remove extra field 'pattern' from PatternApplyForm
The field 'pattern' in PatternApplyForm is no longer needed. Previously,
the field was meant to be instantiated as a hidden field with a default
value equal to the id of the Pattern that the user wanted to apply. This
was a complicated workaround, and was a rather deceptive way of doing
things (i.e. the Pattern to be applied was not determined by the Pattern
indicated in the url but rather by one contained in the POST fields).
This has been changed in favor of extracting the id of the Pattern to
apply from the url.