Rename Pattern to Deriver; use ArabicTools.Pattern
Renames model Dictionary.Pattern to Dictionary.Deriver and changes its
apply() function to rely on the functionality of ArabicTools.Pattern's
generate_spelling() method, as opposed to the method of the same name in
the now-deprecated ArabicTools.AbstractPattern mixin.

NOTE: This breaks basically anything which relies on the Dictionary app,
which is basically the entire project.